+++ src/options.h
@@ -13,7 +13,7 @@ public:
   char *dst_addr;
   char *fname;
   char *av_name;
-  int transfer_size;
+  size_t transfer_size;
   int options;
   hps_rma_opcodes rma_op;
